Move 'ui.widgets' subpackage to 'ui.cell'
