+++ src/edu/colorado/csdms/wmt/client/control/ModelSerializer.java
@@ -37,8 +37,8 @@ import edu.colorado.csdms.wmt.client.data.ModelComponentJSO;
 import edu.colorado.csdms.wmt.client.data.ModelComponentParametersJSO;
 import edu.colorado.csdms.wmt.client.data.ModelJSO;
 import edu.colorado.csdms.wmt.client.ui.ModelTree;
+import edu.colorado.csdms.wmt.client.ui.cell.ComponentCell;
 import edu.colorado.csdms.wmt.client.ui.handler.ComponentSelectionCommand;
-import edu.colorado.csdms.wmt.client.ui.widgets.ComponentCell;
 
 /**
  * Serializes the model built in a WMT session.
