+++ src/edu/colorado/csdms/wmt/client/ui/ParameterTable.java
@@ -32,9 +32,9 @@ import com.google.gwt.user.client.ui.HasHorizontalAlignment;
 
 import edu.colorado.csdms.wmt.client.control.DataManager;
 import edu.colorado.csdms.wmt.client.data.ParameterJSO;
+import edu.colorado.csdms.wmt.client.ui.cell.DescriptionCell;
+import edu.colorado.csdms.wmt.client.ui.cell.ValueCell;
 import edu.colorado.csdms.wmt.client.ui.panel.ParameterActionPanel;
-import edu.colorado.csdms.wmt.client.ui.widgets.DescriptionCell;
-import edu.colorado.csdms.wmt.client.ui.widgets.ValueCell;
 
 /**
  * Builds a table of parameters for a single WMT model component. The value of
