+++ src/edu/colorado/csdms/wmt/client/ui/cell/DoubleCell.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package edu.colorado.csdms.wmt.client.ui.widgets;
+package edu.colorado.csdms.wmt.client.ui.cell;
 
 import java.text.ParseException;
 
@@ -29,44 +29,42 @@ import com.google.gwt.core.client.GWT;
 import com.google.gwt.dom.client.Document;
 import com.google.gwt.event.logical.shared.ValueChangeEvent;
 import com.google.gwt.event.logical.shared.ValueChangeHandler;
-import com.google.gwt.text.client.IntegerParser;
-import com.google.gwt.text.client.IntegerRenderer;
 import com.google.gwt.user.client.ui.ValueBox;
 
 import edu.colorado.csdms.wmt.client.data.ParameterJSO;
 
 /**
- * A ValueBox that uses {@link IntegerParser} and {@link IntegerRenderer} to
- * display an "int" parameter type. {@link ValueChangeEvent}s are sent to
- * {@link IntegerCellChangeHandler}.
+ * A ValueBox that uses {@link DoubleCellParser} and {@link DoubleCellRenderer}.
+ * This is the default for the "float" parameter type. {@link ValueChangeEvent}s
+ * are sent to {@link DoubleCellChangeHandler}.
  * 
  * @author Mark Piper (mark.piper@colorado.edu)
  */
-public class IntegerCell extends ValueBox<Integer> {
+public class DoubleCell extends ValueBox<Double> {
 
   private ValueCell parent;
   private ParameterJSO parameter;
 
   /**
-   * Makes a new {@link IntegerCell} and parses, formats and displays its value.
+   * Makes a new {@link DoubleCell} and parses, formats and displays its value.
    * 
-   * @param parent the parent of the IntegerCell, a ValueCell
+   * @param parent the parent of the DoubleCell, a ValueCell
    */
-  public IntegerCell(ValueCell parent) {
+  public DoubleCell(ValueCell parent) {
 
-    super(Document.get().createTextInputElement(), IntegerRenderer.instance(),
-        IntegerParser.instance());
+    super(Document.get().createTextInputElement(), DoubleCellRenderer
+        .instance(), DoubleCellParser.instance());
 
     this.parent = parent;
     this.parameter = this.parent.getParameter();
 
-    addValueChangeHandler(new IntegerCellChangeHandler());
+    addValueChangeHandler(new DoubleCellChangeHandler());
     setStyleName("wmt-ValueBoxen");
-    
+
     try {
-      Integer integerValue =
-          Integer.valueOf(parameter.getValue().getDefault());
-      setValue(integerValue);
+      Double doubleValue =
+          Double.valueOf(this.parameter.getValue().getDefault());
+      setValue(doubleValue);
       setStyleDependentName("outofrange", !isInRange());
     } catch (Exception e) {
       setValue(null);
@@ -76,21 +74,11 @@ public class IntegerCell extends ValueBox<Integer> {
 
   /**
    * Checks whether a given value is within the established range of values for
-   * the current parameter of the {@link ValueCell}, returning a Boolean. Also
-   * checks that the parameter's min/max values are within the Integer range.
+   * the current parameter of the {@link ValueCell}, returning a Boolean.
    */
   public Boolean isInRange() {
-    Integer minValue, maxValue;
-    try {
-      minValue = Integer.valueOf(parameter.getValue().getMin());
-    } catch (NumberFormatException e) {
-      minValue = Integer.MIN_VALUE;
-    }
-    try {
-      maxValue = Integer.valueOf(parameter.getValue().getMax());
-    } catch (NumberFormatException e) {
-      maxValue = Integer.MAX_VALUE;
-    }
+    Double minValue = Double.valueOf(parameter.getValue().getMin());
+    Double maxValue = Double.valueOf(parameter.getValue().getMax());
     Boolean rangeOK = false;
     if ((getValue() <= maxValue) && (getValue() >= minValue)) {
       rangeOK = true;
@@ -99,17 +87,17 @@ public class IntegerCell extends ValueBox<Integer> {
   }
 
   /**
-   * A class to handle edit events in an {@link IntegerCell}. The
+   * A class to handle edit events in a {@link DoubleCell}. The
    * {@link ValueChangeEvent} is fired when the <code>Enter</code> or
    * <code>Tab</code> keys are pressed, or when focus leaves the
-   * {@link IntegerCell}. Also checks for valid contents of the boxen.
+   * {@link DoubleCell}. Also checks for valid contents of the boxen.
    */
-  public class IntegerCellChangeHandler implements ValueChangeHandler<Integer> {
+  public class DoubleCellChangeHandler implements ValueChangeHandler<Double> {
     @Override
-    public void onValueChange(ValueChangeEvent<Integer> event) {
-      GWT.log("(onValueChange:IntegerCell)");
+    public void onValueChange(ValueChangeEvent<Double> event) {
+      GWT.log("(onValueChange:DoubleCell)");
       try {
-        Integer value = getValueOrThrow();
+        Double value = getValueOrThrow();
         if (value == null) {
           addStyleDependentName("outofrange");
           return;
