+++ src/edu/colorado/csdms/wmt/client/ui/cell/DoubleCellParser.java
@@ -21,9 +21,42 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
+package edu.colorado.csdms.wmt.client.ui.cell;
+
+import java.text.ParseException;
+
+import com.google.gwt.i18n.client.NumberFormat;
+import com.google.gwt.text.shared.Parser;
+
 /**
- * Holds the basic widgets and dialogs that are used to build the WMT GUI.
- * 
- * @author Mark Piper (mark.piper@colorado.edu)
+ * A localized parser based on {@link NumberFormat#getDecimalFormat}.
  */
-package edu.colorado.csdms.wmt.client.ui.widgets;
+public class DoubleCellParser implements Parser<Double> {
+
+  private static DoubleCellParser INSTANCE;
+
+  /**
+   * Returns the instance of the no-op renderer.
+   */
+  public static Parser<Double> instance() {
+    if (INSTANCE == null) {
+      INSTANCE = new DoubleCellParser();
+    }
+    return INSTANCE;
+  }
+
+  protected DoubleCellParser() {
+  }
+
+  public Double parse(CharSequence object) throws ParseException {
+    if ("".equals(object.toString())) {
+      return null;
+    }
+
+    try {
+      return NumberFormat.getDecimalFormat().parse(object.toString());
+    } catch (NumberFormatException e) {
+      throw new ParseException(e.getMessage(), 0);
+    }
+  }
+}
