+++ src/edu/colorado/csdms/wmt/client/ui/cell/IntegerCell.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package edu.colorado.csdms.wmt.client.ui.widgets;
+package edu.colorado.csdms.wmt.client.ui.cell;
 
 import java.text.ParseException;
 
@@ -29,42 +29,44 @@ import com.google.gwt.core.client.GWT;
 import com.google.gwt.dom.client.Document;
 import com.google.gwt.event.logical.shared.ValueChangeEvent;
 import com.google.gwt.event.logical.shared.ValueChangeHandler;
+import com.google.gwt.text.client.IntegerParser;
+import com.google.gwt.text.client.IntegerRenderer;
 import com.google.gwt.user.client.ui.ValueBox;
 
 import edu.colorado.csdms.wmt.client.data.ParameterJSO;
 
 /**
- * A ValueBox that uses {@link DoubleCellParser} and {@link DoubleCellRenderer}.
- * This is the default for the "float" parameter type. {@link ValueChangeEvent}s
- * are sent to {@link DoubleCellChangeHandler}.
+ * A ValueBox that uses {@link IntegerParser} and {@link IntegerRenderer} to
+ * display an "int" parameter type. {@link ValueChangeEvent}s are sent to
+ * {@link IntegerCellChangeHandler}.
  * 
  * @author Mark Piper (mark.piper@colorado.edu)
  */
-public class DoubleCell extends ValueBox<Double> {
+public class IntegerCell extends ValueBox<Integer> {
 
   private ValueCell parent;
   private ParameterJSO parameter;
 
   /**
-   * Makes a new {@link DoubleCell} and parses, formats and displays its value.
+   * Makes a new {@link IntegerCell} and parses, formats and displays its value.
    * 
-   * @param parent the parent of the DoubleCell, a ValueCell
+   * @param parent the parent of the IntegerCell, a ValueCell
    */
-  public DoubleCell(ValueCell parent) {
+  public IntegerCell(ValueCell parent) {
 
-    super(Document.get().createTextInputElement(), DoubleCellRenderer
-        .instance(), DoubleCellParser.instance());
+    super(Document.get().createTextInputElement(), IntegerRenderer.instance(),
+        IntegerParser.instance());
 
     this.parent = parent;
     this.parameter = this.parent.getParameter();
 
-    addValueChangeHandler(new DoubleCellChangeHandler());
+    addValueChangeHandler(new IntegerCellChangeHandler());
     setStyleName("wmt-ValueBoxen");
-
+    
     try {
-      Double doubleValue =
-          Double.valueOf(this.parameter.getValue().getDefault());
-      setValue(doubleValue);
+      Integer integerValue =
+          Integer.valueOf(parameter.getValue().getDefault());
+      setValue(integerValue);
       setStyleDependentName("outofrange", !isInRange());
     } catch (Exception e) {
       setValue(null);
@@ -74,11 +76,21 @@ public class DoubleCell extends ValueBox<Double> {
 
   /**
    * Checks whether a given value is within the established range of values for
-   * the current parameter of the {@link ValueCell}, returning a Boolean.
+   * the current parameter of the {@link ValueCell}, returning a Boolean. Also
+   * checks that the parameter's min/max values are within the Integer range.
    */
   public Boolean isInRange() {
-    Double minValue = Double.valueOf(parameter.getValue().getMin());
-    Double maxValue = Double.valueOf(parameter.getValue().getMax());
+    Integer minValue, maxValue;
+    try {
+      minValue = Integer.valueOf(parameter.getValue().getMin());
+    } catch (NumberFormatException e) {
+      minValue = Integer.MIN_VALUE;
+    }
+    try {
+      maxValue = Integer.valueOf(parameter.getValue().getMax());
+    } catch (NumberFormatException e) {
+      maxValue = Integer.MAX_VALUE;
+    }
     Boolean rangeOK = false;
     if ((getValue() <= maxValue) && (getValue() >= minValue)) {
       rangeOK = true;
@@ -87,17 +99,17 @@ public class DoubleCell extends ValueBox<Double> {
   }
 
   /**
-   * A class to handle edit events in a {@link DoubleCell}. The
+   * A class to handle edit events in an {@link IntegerCell}. The
    * {@link ValueChangeEvent} is fired when the <code>Enter</code> or
    * <code>Tab</code> keys are pressed, or when focus leaves the
-   * {@link DoubleCell}. Also checks for valid contents of the boxen.
+   * {@link IntegerCell}. Also checks for valid contents of the boxen.
    */
-  public class DoubleCellChangeHandler implements ValueChangeHandler<Double> {
+  public class IntegerCellChangeHandler implements ValueChangeHandler<Integer> {
     @Override
-    public void onValueChange(ValueChangeEvent<Double> event) {
-      GWT.log("(onValueChange:DoubleCell)");
+    public void onValueChange(ValueChangeEvent<Integer> event) {
+      GWT.log("(onValueChange:IntegerCell)");
       try {
-        Double value = getValueOrThrow();
+        Integer value = getValueOrThrow();
         if (value == null) {
           addStyleDependentName("outofrange");
           return;
