+++ src/edu/colorado/csdms/wmt/client/ui/cell/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Cells are basic UI elements that hold values in the WMT client interface.
+ * 
+ * @author Mark Piper (mark.piper@colorado.edu)
+ */
+package edu.colorado.csdms.wmt.client.ui.cell;
