+++ src/edu/colorado/csdms/wmt/client/ui/handler/ComponentGetInformationCommand.java
@@ -27,8 +27,8 @@ import com.google.gwt.core.client.GWT;
 import com.google.gwt.user.client.Command;
 
 import edu.colorado.csdms.wmt.client.control.DataManager;
+import edu.colorado.csdms.wmt.client.ui.cell.ComponentCell;
 import edu.colorado.csdms.wmt.client.ui.dialog.ComponentInfoDialogBox;
-import edu.colorado.csdms.wmt.client.ui.widgets.ComponentCell;
 
 /**
  * Defines the action for the "Get info" menu item in a {@link ComponentCell};
