+++ src/edu/colorado/csdms/wmt/client/ui/handler/ComponentSelectionCommand.java
@@ -28,8 +28,8 @@ import com.google.gwt.user.client.Command;
 
 import edu.colorado.csdms.wmt.client.Constants;
 import edu.colorado.csdms.wmt.client.control.DataManager;
+import edu.colorado.csdms.wmt.client.ui.cell.ComponentCell;
 import edu.colorado.csdms.wmt.client.ui.menu.ComponentActionMenu;
-import edu.colorado.csdms.wmt.client.ui.widgets.ComponentCell;
 
 /**
  * Defines the action for when a user selects a component in a
