+++ src/edu/colorado/csdms/wmt/client/ui/handler/ComponentShowParametersCommand.java
@@ -27,7 +27,7 @@ import com.google.gwt.core.shared.GWT;
 import com.google.gwt.user.client.Command;
 
 import edu.colorado.csdms.wmt.client.control.DataManager;
-import edu.colorado.csdms.wmt.client.ui.widgets.ComponentCell;
+import edu.colorado.csdms.wmt.client.ui.cell.ComponentCell;
 
 /**
  * Defines the action for the "Show parameters" menu item in a
